Load the esbuild bundle backend lazily so cloud bundles stay Workers-safe
